% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{koon_browsers}
\alias{koon_browsers}
\title{List all available browser profile names.}
\usage{
koon_browsers()
}
\value{
A character vector of browser profile names.
}
\description{
List all available browser profile names.
}
