% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R, R/koon.R
\docType{data}
\name{koon_browsers}
\alias{koon_browsers}
\title{List all available browser profile names.}
\format{
An object of class \code{NULL} of length 0.

An object of class \code{NULL} of length 0.
}
\usage{
koon_browsers

koon_browsers
}
\value{
A character vector of browser profile names.

A character vector of browser profile names that can be passed
to \code{Koon$new()}.
}
\description{
List all available browser profile names.

List All Available Browser Profiles
}
\examples{
\dontrun{
koon_browsers()
# [1] "chrome131" "chrome132" ... "opera127"
}

}
\keyword{datasets}
